Add files via upload
*** IN THE LIBRARIAN CLASS, I added a bunch of notes and updates to the select functions/views; if they don't make sense (probably won't), please let me know, and we can chat

LMS - just commented out the main method, so it's a useless shell at the moment.

The print Results for the query right now is an ugly console print. I need to do some more thinking about how to print to windows - I think we made each window a class...so that may be a lot of classes for us... ... ... but i wanted to give you something that you could use to start testing / printing. and it's in array form, so it's ready for the window...i just have to figure out how to do that.